requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Admin==1.6.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Werkzeug==2.3.7
email-validator==2.1.0
Pillow==10.2.0
Flask-JWT-Extended==4.6.0
marshmallow==3.21.0
flask-marshmallow==1.2.1
marshmallow-sqlalchemy==1.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
openpyxl==3.1.2
reportlab==4.1.0
pyotp==2.9.0
qrcode[pil]==7.4.2
Flask-Mail==0.10.0
